requirements.txt
numpy>=1.24.0
pandas>=2.0.0
label-studio>=1.9.0
label-studio-sdk>=0.0.32
torch>=2.0.0
transformers>=4.30.0
spacy>=3.6.0
scikit-learn>=1.3.0
datasets>=2.14.0
jsonlines>=4.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
tabulate>=0.9.0
anthropic>=0.34.0
openai>=1.0.0
